refactor(run): Move the `runner` interface to its own file
