+++ cmd/kraft/run/run.go
@@ -5,7 +5,6 @@
 package run
 
 import (
-	"context"
 	"fmt"
 	"strings"
 
@@ -91,43 +90,6 @@ func New() *cobra.Command {
 	)
 
 	return cmd
-}
-
-// runner is an interface for defining different mechanisms to execute the
-// provided unikernel.  Standardizing first the check, Runnable, to determine
-// whether the provided input is capable of executing, and Prepare, actually
-// performing the preparation of the Machine specification for the controller.
-type runner interface {
-	// String implements fmt.Stringer and returns the name of the implementing
-	// runner.
-	fmt.Stringer
-
-	// Runnable checks whether the provided configuration is runnable.
-	Runnable(context.Context, *Run, ...string) (bool, error)
-
-	// Prepare the provided configuration into a machine specification ready for
-	// execution by the controller.
-	Prepare(context.Context, *Run, *machineapi.Machine, ...string) error
-}
-
-// runners is the list of built-in runners which are checked sequentially for
-// capability.  The first to test positive via Runnable is used with the
-// controller.
-func runners() map[string]runner {
-	r := map[string]runner{
-		"linuxu":  &runnerLinuxu{},
-		"kernel":  &runnerKernel{},
-		"project": &runnerProject{},
-		"package": &runnerPackage{},
-	}
-
-	for k, pm := range packmanager.PackageManagers() {
-		r[string(k)] = &runnerPackage{
-			pm: pm,
-		}
-	}
-
-	return r
 }
 
 func (opts *Run) Pre(cmd *cobra.Command, _ []string) error {
